REFACTOR: changed name of Product to ExtensibleProduct, hopefully one last time
