+++ src/components/ExtensibleProduct.jsx
